Use '--no-pager' flag for git-log (#228)
When there are a log of changes in the git step, topgrade might be "stack" in pager (e.g. less) until the user will manually exit it.
Using '--no-pager' flag allow all the changes to be printed to screen regardless of the amount of commits or the user's git configuration.