Linux: System: Fix rpm-ostree detection (#816)
On a rpm-ostree based system, topgrade would previously fail to detect
the rpm-ostree executable and fall back to yum instead, which isn't
available. This was caused by constructing a Path instance with
`Path::new`, rather than querying the underlying OS for a path to the
rpm-ostree executable.

Make the `system` update step use `which` to determine if an executable
called "rpm-ostree" is available on the system and get the correct path
to the executable.